+++ packages/services/src/Domain/Item/ItemCounterInterface.ts
@@ -1,0 +1,5 @@
+import { SNNote, SNTag, ItemCounts } from '@standardnotes/models'
+
+export interface ItemCounterInterface {
+  countNotesAndTags(items: Array<SNNote | SNTag>): ItemCounts
+}