+++ packages/api/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.8](https://github.com/standardnotes/app/compare/@standardnotes/api@1.4.7...@standardnotes/api@1.4.8) (2022-08-23)
+
+**Note:** Version bump only for package @standardnotes/api
+
 ## [1.4.7](https://github.com/standardnotes/app/compare/@standardnotes/api@1.4.6...@standardnotes/api@1.4.7) (2022-08-05)
 
 **Note:** Version bump only for package @standardnotes/api