+++ packages/files/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.1](https://github.com/standardnotes/app/compare/@standardnotes/files@1.9.0...@standardnotes/files@1.9.1) (2022-08-23)
+
+**Note:** Version bump only for package @standardnotes/files
+
 # [1.9.0](https://github.com/standardnotes/app/compare/@standardnotes/files@1.8.7...@standardnotes/files@1.9.0) (2022-08-05)
 
 ### Bug Fixes