+++ packages/mobile/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.31.16](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.31.15...@standardnotes/mobile@3.31.16) (2022-08-23)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
 ## [3.31.15](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.31.14...@standardnotes/mobile@3.31.15) (2022-08-22)
 
 **Note:** Version bump only for package @standardnotes/mobile